+++ src/connection/notify.rs
@@ -13,8 +13,8 @@ impl Notify {
     /**
      * process ID of notifying server process
      */
-    pub fn be_pid(&self) -> i32 {
-        unsafe { (*self.notify).be_pid }
+    pub fn be_pid(&self) -> u32 {
+        unsafe { (*self.notify).be_pid as u32 }
     }
 
     /**
